找到字符串中所有字母异位词,和为 K 的子数组

+++ include/header.hpp
@@ -9,6 +9,7 @@
 #pragma once
 
 #include <algorithm>
+#include <cstddef>
 #include <vector>
 #include <string>
 #include <iterator>
@@ -44,7 +45,7 @@ static ListNode *GenList(vector<int> arr) {
     ListNode *head = NULL;
     ListNode *node = NULL;
 
-    for (int i = 0; i < arr.size(); i++) {
+    for (size_t i = 0; i < arr.size(); i++) {
         ListNode *ln = new ListNode(arr[i]);
 
         if (!node) {
